+++ great_expectations/data_context/store/html_site_store.py
@@ -1,5 +1,6 @@
 import logging
 import os
+import inspect
 from mimetypes import guess_type
 
 from great_expectations.data_context.types.resource_identifiers import (